+++ cli/integration-test/integration/fixture.clj
@@ -12,7 +12,7 @@
      :params params
      :id (lsp/inc-request-id)}))
 
-(def default-init-options {:lint-project-files-after-startup? false})
+(def default-init-options {})
 
 (defn initialize-request
   ([]