+++ lib/src/clojure_lsp/diff.clj
@@ -40,22 +40,12 @@
                                 (shared/relativize-filepath project-path)))]
          unlines)))
 
-(def ^:private ansi-colors
-  {:reset "[0m"
-   :red   "[31m"
-   :green "[32m"
-   :yellow "[33m"
-   :cyan  "[36m"})
-
-(defn colorize [s color]
-  (str \u001b (ansi-colors color) s \u001b (ansi-colors :reset)))
-
 (defn colorize-diff [diff-text]
   (-> diff-text
-      (str/replace #"(?m)^(rename from .*)$"  (colorize "$1" :yellow))
-      (str/replace #"(?m)^(rename to .*)$"  (colorize "$1" :yellow))
-      (str/replace #"(?m)^(\-\-\-\sa.*)$"  (colorize "$1" :yellow))
-      (str/replace #"(?m)^(\+\+\+\sb.*)$"  (colorize "$1" :yellow))
-      (str/replace #"(?m)^(@@.*@@)$"       (colorize "$1" :cyan))
-      (str/replace #"(?m)^(\+(?!\+\+).*)$" (colorize "$1" :green))
-      (str/replace #"(?m)^(-(?!--).*)$"    (colorize "$1" :red))))
+      (str/replace #"(?m)^(rename from .*)$"  (shared/colorize "$1" :yellow))
+      (str/replace #"(?m)^(rename to .*)$"  (shared/colorize "$1" :yellow))
+      (str/replace #"(?m)^(\-\-\-\sa.*)$"  (shared/colorize "$1" :yellow))
+      (str/replace #"(?m)^(\+\+\+\sb.*)$"  (shared/colorize "$1" :yellow))
+      (str/replace #"(?m)^(@@.*@@)$"       (shared/colorize "$1" :cyan))
+      (str/replace #"(?m)^(\+(?!\+\+).*)$" (shared/colorize "$1" :green))
+      (str/replace #"(?m)^(-(?!--).*)$"    (shared/colorize "$1" :red))))