+++ lib/src/clojure_lsp/feature/refactor.clj
@@ -16,17 +16,17 @@
 
 (defmulti refactor :refactoring)
 
-(defmethod refactor :add-import-to-namespace [{:keys [loc args] {:keys [db]} :components}]
-  (apply f.add-missing-libspec/add-import-to-namespace loc (concat args [db])))
+(defmethod refactor :add-import-to-namespace [{:keys [loc uri args] {:keys [db]} :components}]
+  (apply f.add-missing-libspec/add-missing-import loc uri (concat args [db])))
 
 (defmethod refactor :add-missing-libspec [{:keys [loc uri] {:keys [db]} :components}]
   (f.add-missing-libspec/add-missing-libspec loc uri db))
 
-(defmethod refactor :add-require-suggestion [{:keys [loc args] {:keys [db]} :components}]
-  (apply f.add-missing-libspec/add-require-suggestion loc (concat args [db])))
+(defmethod refactor :add-require-suggestion [{:keys [loc uri args] {:keys [db]} :components}]
+  (apply f.add-missing-libspec/add-require-suggestion loc uri (concat args [db])))
 
-(defmethod refactor :add-missing-import [{:keys [loc] {:keys [db]} :components}]
-  (f.add-missing-libspec/add-common-import-to-namespace loc db))
+(defmethod refactor :add-missing-import [{:keys [loc uri] {:keys [db]} :components}]
+  (f.add-missing-libspec/add-missing-import loc uri nil db))
 
 (defmethod refactor :clean-ns [{:keys [loc uri] {:keys [db]} :components}]
   (f.clean-ns/clean-ns-edits loc uri db))