+++ lib/src/clojure_lsp/internal_api.clj
@@ -41,15 +41,15 @@
   (-info [_ arg1] (log-print "[INFO]" options arg1))
   (-info [_ arg1 arg2] (log-print "[INFO]" options arg1 arg2))
   (-info [_ arg1 arg2 arg3] (log-print "[INFO]" options arg1 arg2 arg3))
-  (-warn [_ arg1] (log-print "[WARN]" options arg1))
-  (-warn [_ arg1 arg2] (log-print "[WARN]" options arg1 arg2))
-  (-warn [_ arg1 arg2 arg3] (log-print "[WARN]" options arg1 arg2 arg3))
-  (-error [_ arg1] (log-print "[ERROR]" options arg1))
-  (-error [_ arg1 arg2] (log-print "[ERROR]" options arg1 arg2))
-  (-error [_ arg1 arg2 arg3] (log-print "[ERROR]" options arg1 arg2 arg3))
-  (-debug [_ arg1] (log-print "[DEBUG]" options arg1))
-  (-debug [_ arg1 arg2] (log-print "[DEBUG]" options arg1 arg2))
-  (-debug [_ arg1 arg2 arg3] (log-print "[DEBUG]" options arg1 arg2 arg3)))
+  (-warn [_ arg1] (log-print (shared/colorize "[WARN]" :yellow) options arg1))
+  (-warn [_ arg1 arg2] (log-print (shared/colorize "[WARN]" :yellow) options arg1 arg2))
+  (-warn [_ arg1 arg2 arg3] (log-print (shared/colorize "[WARN]" :yellow) options arg1 arg2 arg3))
+  (-error [_ arg1] (log-print (shared/colorize "[ERROR]" :red) options arg1))
+  (-error [_ arg1 arg2] (log-print (shared/colorize "[ERROR]" :red) options arg1 arg2))
+  (-error [_ arg1 arg2 arg3] (log-print (shared/colorize "[ERROR]" :red) options arg1 arg2 arg3))
+  (-debug [_ arg1] (log-print (shared/colorize "[DEBUG]" :cyan) options arg1))
+  (-debug [_ arg1 arg2] (log-print (shared/colorize "[DEBUG]" :cyan) options arg1 arg2))
+  (-debug [_ arg1 arg2 arg3] (log-print (shared/colorize "[DEBUG]" :cyan) options arg1 arg2 arg3)))
 
 (defn ^:private show-message-cli [options {:keys [message extra type]}]
   (cli-println options (format "\n[%s] %s" (string/upper-case (name type)) message))
@@ -168,8 +168,7 @@
       {:workspace {:workspace-edit {:document-changes true}}}
       (settings/clean-client-settings {})
       (merge (shared/assoc-some
-               {:lint-project-files-after-startup? false
-                :text-document-sync-kind :full}
+               {:text-document-sync-kind :full}
                :log-path log-path)
              settings)
       nil
@@ -310,7 +309,7 @@
                                        (name (f.diagnostic/severity->level severity))
                                        code
                                        message)
-                         (not raw?) (diff/colorize (f.diagnostic/severity->color severity))))
+                         (not raw?) (shared/colorize (f.diagnostic/severity->color severity))))
                      diags)))
             diagnostics)))